+++ test-resources/ns1.clj
@@ -14,11 +14,13 @@
    (clojure data edn)
    [clojure.pprint :refer [get-pretty-writer formatter cl-format]]
    clojure.test.junit
-   [clojure.xml])
+   [clojure.xml]
+   [defines-deftype])
   (:use clojure.test
         clojure.test
         [clojure.string :rename {replace foo reverse bar} :reload-all true :reload true])
-  (:import java.util.Random
+  (:import [defines_deftype SomeDefType]
+           java.util.Random
            java.io.PushbackReader
            java.io.PushbackReader
            java.io.FilenameFilter
@@ -29,6 +31,8 @@
             SomeClass$InnerClass$InnerInnerClassTwo
             SomeClass$InnerClass$InnerInnerClassThree]
            (java.util Date Calendar)))
+
+(SomeDefType.)
 
 (defmacro tt [writer]
   (Random.)