+++ test-resources/bar/ns/libspecs.clj
@@ -1,0 +1,1 @@
+(ns bar.ns.libspecs)