+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* [#142](https://github.com/clojure-emacs/refactor-nrepl/issues/142): Guard `read-ns-form` against non-existing files.
+
 ## 3.3.1
 
 * [#363](https://github.com/clojure-emacs/refactor-nrepl/issues/363): Fix a memoization bug in `clean-namespace`.